requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
docx2pdf==0.1.8
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
